Separated the code so some of it would work
Will have to pick and choose what we want will not need all of it.
