+++ src/ClientConnection.java
@@ -1,0 +1,23 @@
+import java.io.IOException;
+import java.io.OutputStream;
+import java.io.PrintWriter;
+import java.net.Socket;
+
+public class ClientConnection {
+
+    private Socket socket;
+    private PrintWriter out;
+
+    public ClientConnection(Socket socket) throws IOException {
+        this.socket = socket;
+        this.out = new PrintWriter(socket.getOutputStream(), true);
+    }
+
+    public void sendMessage(String message) {
+        out.println(message);  // Send message to client
+    }
+
+    public String getAddress() {
+        return socket.getInetAddress().toString();  // Get client address
+    }
+}
